owner neglected in assigning workspace TC and also refactored with helper function
