+++ apps/teams/forms.py
@@ -4,7 +4,7 @@ from apps.organizations.models import OrganizationMember
 from apps.workspaces.selectors import get_organization_members_by_organization_id
 from apps.teams.models import TeamMember
 from apps.teams.constants import TeamMemberRole
-
+from apps.core.selectors import get_org_members_without_owner
 
 class TeamForm(forms.ModelForm):
     team_coordinator = forms.ModelChoiceField(
@@ -47,9 +47,7 @@ class TeamForm(forms.ModelForm):
         if self.organization:
             self.fields[
                 "team_coordinator"
-            ].queryset = get_organization_members_by_organization_id(
-                self.organization.organization_id
-            )
+            ].queryset = get_org_members_without_owner(self.organization)
         if not self.can_change_team_coordinator:
             self.fields["team_coordinator"].widget.attrs["disabled"] = True
